Combine empoyers- and employees -tables into users -table
Change the overly complicated database structure where employers and
employees are separated into different tables into one where both are
in the users -table and the column "role" tells whether the user is an
employer or an employee. After this change, one can insert and select
information related to a user without checking the role every time.
In the previous structure one had to check the role to know whether to
manipulate the employees are the employers -table.

Other consequences of the change:
- the job table doesn't include information on the selected employee
	- same information is available in the applications -table by
	combining information on the status of the application and the
	job id
- there will likely be no functionality for adding a secured job into the
job experience -part of one's profile directly