Split the get_job_info -function in job.py into two distinct functions
Previously the function returned both the information on the job and on
whether the user had applied for the given job. Now get_job_info in jobs.py
returns only the information on the job while applied_or_not in applications.py
returns info on whether the user has applied for the given job. This allows
getting rid of the complicated left join structure in the sql query.