Ensure that the applicant can apply only once and allow observing the employer's profile
To ensure that the applicant can apply for a  specific job only once, I splitted the function
"get_open_jobs" into employer and employee specific functions. The employee -version of the
function lists all open positions.The employer version lists only such jobs that the applicant has not applied for.
This together with adding a unique -constraint on the applications -table ensures that the  applicant
can't apply for the same job multiple times.

Additionally, I splitted the "/show_profile" -route into applicant and employer specific
routes. These routes receive different parameters and they pass different parameters to the html -file.
"applicant = True/False" -parameter is used for telling the html -file from which route the user
is coming from. This allows for different pages that one can return to after observing the profile.