Fix problems in returning the open jobs for an employee
The function "get_open_jobs_employee" in jobs.py didn't return all such jobs that the
employee hadn't applied for and for which the application period was still
ongoing. This was due to the column u.id referring to the employees rather
than the employee in the sql query. Now the id of the employee is given to the function
as a parameter.