Create a possibility to delete a profile text
Additionally, change the way the editing of a profile text
is handled. Previously the old  profile texts were not deleted
but the column "visible" was set to 0 when a new profile text
was added. Because other tables include no references to the profile
texts, the profile texts can be deleted with a "DELETE"
statement as well. Additionally, adding a new profile text and deleting
the old one are handled by separate functions.

Additionally, code quality has been improved.